Matrix: Modified the code in RawDataTools::addMasterOptionsDataToVector() to (hopefully) make it clearer what's going on with the data bytes for basic channel and the "unused" parameters. Sorted out several problems with the checksum calculation: 1) I was adding the pedal1ControllerNumber value to the checksum again instead of the pedal2ControllerNumber; 2) the value for the unused velocity sensitivity parameter needs to be added; 3) the basic channel displayed value needs to be added. I noticed that the display brightness value can be changed from the plugin, but the actual brightness only changes when you set the value on the hardware itself.

+++ Matrix-6G/Matrix-6G/Source/midi/midi_RawDataTools.h
@@ -51,7 +51,7 @@ struct RawDataTools {
 
 private:
     static const uint8 indexofFirstPatchOrSplitParamDataLSByte{ 20 };
-    static const uint8 indexOfFirstMasterOptionDataLSByte{ 6 };
+    static const uint16 indexOfFirstMasterOptionDataLSByte{ 6 };
     static const int negativeValueOffset{ 256 };
 
     static void applyPatchNumberToGUI(const uint8 patchNumber, UnexposedParameters* unexposedParams);
